Move Camera struct to new file (camera.rs)
