+++ src/camera.rs
@@ -1,0 +1,80 @@
+use vector::{Vector3, Vector2};
+
+#[derive(Debug)]
+pub struct Camera {
+    pub eye: Vector3,
+    pub forward: Vector3,
+    pub right: Vector3,
+    pub up: Vector3,
+    pub zoom: f64,
+}
+
+#[derive(Clone, Debug)]
+pub struct Ray {
+    pub origin: Vector3,
+    pub direction: Vector3,
+}
+
+impl Camera {
+    pub fn new(eye: Vector3, target: Vector3, y_up: Vector3, zoom: f64) -> Camera {
+        let forward = (target - eye).normalize();
+        let right = forward.cross(&y_up).normalize();
+
+        Camera {
+            eye: eye,
+            forward: forward,
+            right: right,
+            up: right.cross(&forward).normalize(),
+            zoom: zoom,
+        }
+    }
+
+    pub fn ray(&self, normalized_coord: &Vector2) -> Ray {
+        Ray {
+            origin: self.eye,
+            direction: (normalized_coord.x * self.right + normalized_coord.y * self.up + self.zoom * self.forward).normalize(),
+        }
+    }
+}
+
+pub struct CameraBuilder {
+    eye: Vector3,
+    target: Vector3,
+    y_up: Vector3,
+    zoom: f64,
+}
+
+impl CameraBuilder {
+    pub fn new() -> CameraBuilder {
+        CameraBuilder {
+            eye: Vector3::zero(),
+            target: Vector3::new(0.0, 0.0, 1.0),
+            y_up: Vector3::new(0.0, 1.0, 0.0),
+            zoom: 2.0,
+        }
+    }
+
+    pub fn eye(&mut self, coordinate: Vector3) -> &mut CameraBuilder {
+        self.eye = coordinate;
+        self
+    }
+
+    pub fn target(&mut self, coordinate: Vector3) -> &mut CameraBuilder {
+        self.target = coordinate;
+        self
+    }
+
+    pub fn y_up(&mut self, coordinate: Vector3) -> &mut CameraBuilder {
+        self.y_up = coordinate;
+        self
+    }
+
+    pub fn zoom(&mut self, coordinate: f64) -> &mut CameraBuilder {
+        self.zoom = coordinate;
+        self
+    }
+
+    pub fn finalize(&self) -> Camera {
+        Camera::new(self.eye, self.target, self.y_up, self.zoom)
+    }
+}
